Add built-in column to node grid
The built-in column does a best-effort attempt to show whether a node
is originally from Dynamo or not. The original idea was to provide
information about which nodes are from the host application, but we
don't have a reliable way to check that. We might want to find a more
realiable way to identify them in the future.

Also, the layout of the row details panel is adjusted to wrap content
so that the checkboxes don't cause horizontal scrolling.